# Implement patch-based source-map concat strategy