--- a/lib/utils/count-new-lines.js
+++ b/lib/utils/count-new-lines.js
@@ -1,0 +1,8 @@
+module.exports = function countNewLines(src) {
+  var newlinePattern = /(\r?\n)/g;
+  var count = 0;
+  while (newlinePattern.exec(src)) {
+    count++;
+  }
+  return count;
+};