--- a/lib/utils/ensure-posix-eol.js
+++ b/lib/utils/ensure-posix-eol.js
@@ -1,0 +1,9 @@
+var EOL = require('os').EOL;
+
+module.exports = function ensurePosixEOL(string) {
+  if (EOL !== '\n') {
+    string = string.split(EOL).join('\n');
+  }
+
+  return string;
+};