… test/strategies/sourcemap-test.js