**Ready for review.**

Similar to https://github.com/broccolijs/broccoli-concat/pull/102, but this time with support for source-maps.

TODO:
- [x] Support concatenation of sourcemaps (assimilate them)
- [x] ~~Add caching of sourcemap encoder/decoder~~ (will do in follow up)
- [x] Add additional tests
  - [x] Unit tests for new SourceMapConcat class
  - [x] Header/Footer multiline scenarios in existing tests
  - [x] ~~Additional rebuild tests~~ (not needed, covered by unit tests above)
- [x] Cleanup code
- [x] Add any new documentation
- [x] Run benchmarks