@trentmwillis 2017-02-27
Ran some benchmarks:

```bash
FILE_COUNT=1000 DEPTH=3 yarn run bench

# new approach
Average time for build over 10 iterations: 426.2ms
Average time for rebuild over 10 iterations: 101.1ms

# old approach
Average time for build over 10 iterations: 353.1ms
Average time for rebuild over 10 iterations: 332.5ms

FILE_COUNT=10 LARGE_FILES=true yarn run bench

# new approach
Average time for build over 10 iterations: 227.4ms
Average time for rebuild over 10 iterations: 196.4ms

# old approach
Average time for build over 10 iterations: 174.9ms
Average time for rebuild over 10 iterations: 163.3ms
```

So similarly to `SimpleConcat`, initial build is slightly slower, but rebuilds are faster for most cases. However, when dealing with large files, there is pretty much no improvement. This is due to a lot of time being spent in creating the sourcemaps. I'm going to run this against our app to see what a real world example looks like.