@rwjblue 2017-02-27
LGTM, let me know once your real world app tests are confirmed working, and we can land + release.