@trentmwillis 2017-02-28
Tested this in our app and Travis-Web. Improvements aren't as good as expected, but the pending change-tracking work will make this much better.

That said, I did catch some bugs during my investigation, so I need to write some additional tests and land them prior to these changes.