@trentmwillis 2017-02-28
Added some additional rebuild tests. This should be good to land whenever, though we should note that integration with Ember-CLI apps is likely to not see a huge win until change-tracking is supported.