@trentmwillis 2017-06-28
Unsure. The last I remember we were waiting for a change-tracking approach for Broccoli to be agreed upon. This current implementation isn't a huge improvement, but it would be able to take advantage of change-tracking if it is implemented.