@trentmwillis 2017-12-06
This seems to have stalled. I'm going to close, but I'll leave the branch intact in case it is decided to pick this up again in the future.